Primer versión funcional del app qrsocial

+++ application/models/cammodel.php
@@ -12,11 +12,39 @@
     function loadcampania($idusuario , $idcuentaactual){
           
         $this->db->where('idcuenta', $idcuentaactual);
-        $this->db->where('status', '1' );
+        
         $this->db->where('idusuario', $idusuario);
         $query = $this->db->get('campaña');      
         
         return $query->result_array();
+
+    }
+    function loadlastcampania($idusuario , $idcuentaactual){
+
+        $query = $this->db->query("select idcampaña from campaña WHERE idcuenta='".$idcuentaactual."'  ORDER BY idcampaña DESC LIMIT 1");      
+        
+        return $query->result_array();
+
+    }
+
+    function ischeckcampzona($zona , $camp ){
+
+        $this->db->where('idzona', $zona); 
+        $this->db->where('idcampaña', $camp); 
+        $query = $this->db->get('campaña_zona');      
+        $n = $query->num_rows();
+        return $n;
+        
+
+    }
+
+    function removecampzona( $zona , $camp ){
+
+        $this->db->where('idzona', $zona); 
+        $this->db->where('idcampaña', $camp); 
+        $query = $this->db->delete('campaña_zona');
+        return $query;        
+
 
     }
 
@@ -31,24 +59,45 @@
 
     function deletecamobyid($id, $idusuario){
 
-        
         $this->db->where('idcampaña', $id);
-        $this->db->where('idusuario', $idusuario );
-        $query_del =$this->db->delete('campaña'); 
+        $query_delmensaje = $this->db->delete('mensaje'); 
+
+
+        if ($query_delmensaje ==  true) {
+
+
+
+                    $this->db->where('idcampaña', $id);
+                    $query_delcampzona = $this->db->delete('campaña_zona'); 
+                    
+
+
+                    $this->db->where('idcampaña', $id);
+                    $this->db->where('idusuario', $idusuario );
+                    $query_del =$this->db->delete('campaña'); 
+                    
+                      if ($query_del ==  true){
+                          return "1";
+                        }else{
+                          return "Falla al eliminar en la base de datos consulte al administrador";
+                        }
+
+          
+        } else {
+                    return "Falla al eliminar en la base de datos consulte al administrador";
+        }
         
-          if ($query_del ==  true){
-              return "1";
-            }else{
-              return "Falla al eliminar en la base de datos consulte al administrador";
-            }
+
+
+
     }
 
-      
-    function editcamp($id , $nombre, $descripcion){
+    function editcamp($id , $nombre, $descripcion ,  $estadocamp){
 
            $data = array(               
                'nombre' => $nombre,
-               'descripcion' => $descripcion
+               'descripcion' => $descripcion, 
+               'status' => $estadocamp
             );
 
             $this->db->where('idcampaña', $id);
@@ -60,6 +109,15 @@
 
 
         $this->db->select('nombre');
+        $this->db->where('idcampaña', $id );
+        $this->db->where('idusuario', $idusuario);  
+        $query = $this->db->get('campaña');
+        return $query->result_array();
+
+    }
+    function getstatusnamecampbyid($id, $idusuario){
+        
+        $this->db->select('status');
         $this->db->where('idcampaña', $id );
         $this->db->where('idusuario', $idusuario);  
         $query = $this->db->get('campaña');
@@ -79,6 +137,16 @@
     }
 
 
+    function registrocampzona($idzona , $idcamp ){
+
+          $data = array(                 
+                 'idcampaña' => $idcamp,
+                 'idzona' => $idzona
+            );
+        $result = $this->db->insert('campaña_zona', $data); 
+
+        return $result;
+   } 
 
     function isnameexist($name , $idusuario ){
 
@@ -92,29 +160,28 @@
 
     function registrocamp($name , $descripcion, $evento , $idusuario , $cuenta){
 
+
+
       $datastatus="";
       if ($this->isnameexist($name , $idusuario) >0){
-            $datastatus ="En el sistema ya existe una campaña con ese nombre, intente con otro";
+            $datastatus ="3";
       }else{
 
             $status = 1; 
             $data = array(
-             'nombre' => $name ,
-             'descripcion' => $descripcion ,         
-             'status' => $status,
-             'evento' => $evento,
-             'idusuario' =>$idusuario, 
-             'idcuenta' => $cuenta
+               'nombre' => $name ,
+               'descripcion' => $descripcion ,         
+               'status' => $status,
+               'evento' => $evento,
+               'idusuario' =>$idusuario, 
+               'idcuenta' => $cuenta
               );
 
              $result = $this->db->insert('campaña', $data);
               if ($result == true) {
                 
-                $datastatus ="1";  
 
-              }else{
-
-                $datastatus ="0";
+                $datastatus ="ok";  
 
               }
        }
